+++ src/main/java/ru/practicum/shareit/item/model/Item.java
@@ -1,6 +1,18 @@
 package ru.practicum.shareit.item.model;
 
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
 
+@Getter
+@Setter
+@NoArgsConstructor
+@AllArgsConstructor
 public class Item {
-
+    private long id;
+    private String name;
+    private String description;
+    private Boolean available;
+    private long owner;
 }
