+++ src/main/java/ru/practicum/shareit/item/service/ItemService.java
@@ -1,7 +1,7 @@
-package ru.practicum.shareit.user.controller.item.service;
+package ru.practicum.shareit.item.service;
 
-import ru.practicum.shareit.user.controller.item.dto.ItemDto;
-import ru.practicum.shareit.user.controller.item.model.Item;
+import ru.practicum.shareit.item.dto.ItemDto;
+import ru.practicum.shareit.item.model.Item;
 
 import java.util.List;
