--- src/main/java/ru/practicum/shareit/user/User.java
@@ -1,6 +1,0 @@
-package ru.practicum.shareit.user;
-
-
-public class User {
-
-}
