--- src/main/java/ru/practicum/shareit/user/UserController.java
@@ -1,11 +1,0 @@
-package ru.practicum.shareit.user;
-
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
-
-
-@RestController
-@RequestMapping(path = "/users")
-public class UserController {
-
-}
