--- src/main/java/ru/practicum/shareit/user/controller/item/dto/ItemDto.java
@@ -1,24 +1,0 @@
-package ru.practicum.shareit.user.controller.item.dto;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-
-import javax.validation.constraints.NotBlank;
-import javax.validation.constraints.NotNull;
-
-@Getter
-@Setter
-@AllArgsConstructor
-@NoArgsConstructor
-public class ItemDto {
-    private long id;
-    @NotBlank
-    private String name;
-    @NotBlank
-    private String description;
-    @NotNull
-    private Boolean available;
-
-}
