--- src/main/java/ru/practicum/shareit/user/controller/item/dto/dto/ItemRequestDto.java
@@ -1,5 +1,0 @@
-package ru.practicum.shareit.user.controller.item.dto.dto;
-
-
-public class ItemRequestDto {
-}
