--- src/main/java/ru/practicum/shareit/user/controller/item/model/Item.java
@@ -1,19 +1,0 @@
-package ru.practicum.shareit.user.controller.item.model;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-
-@Getter
-@Setter
-@NoArgsConstructor
-@AllArgsConstructor
-public class Item {
-    private long id;
-    private String name;
-    private String description;
-    private Boolean available;
-    private long owner;
-
-}
